created policies for book and post

+++ src/app/Http/Controllers/PostsController.php
@@ -11,10 +11,12 @@ use App\Models\Image;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Str;
 use App\Models\User;
+use App\Policies\PostPolicy;
+use Illuminate\Auth\Access\Gate;
+use Illuminate\Support\Facades\Auth;
 
 class PostsController extends Controller
 {
-
     public function filterByCategory(Category $category)
     {
         $posts = $category->posts;
@@ -27,6 +29,9 @@ class PostsController extends Controller
     }
 
     public function uploadImage(Request $request) {
+
+        $this->authorize('uploadImage', Post::class);
+
         $file = $request->file('upload');
 
         $path = $file->store('yolo', 'public');
@@ -55,12 +60,6 @@ class PostsController extends Controller
      */
     public function index(Request $request)
     {
-        // if (auth()->check()) {
-        //     $posts = Post::withTrashed()->with('category', 'images')->get();
-        // } else {
-        //     $posts = Post::with('category', 'images')->get();
-        // }
-
         $posts = Cache::rememberForever('posts.index', function () {
             return Post::withTrashed()->with('category', 'images')->get();
         });
@@ -80,6 +79,8 @@ class PostsController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Post::class);
+
         $categories = Category::orderBy('categoryName')->get();
 
         // Clear the session.
@@ -96,10 +97,12 @@ class PostsController extends Controller
      */
     public function store(StoreBlogPost $request, User $user)
     {
+        $this->authorize('store', Post::class);
+
         $user->with('CRUD posts');
         $post = new Post();
         $post->title = $request->input('title');
-        $post->body = strip_tags($request->input('body'));
+        $post->body = ($request->input('body'));
         
         $post->category_id = $request->input('category_id');
         $post->slug = Str::slug($post->title);
@@ -112,8 +115,6 @@ class PostsController extends Controller
             // Clear the session.
             session()->put('images', []);
         }
-        
-        
         return redirect()->route('posts.index')->with('success', 'Post Created');
     }
 
@@ -139,12 +140,12 @@ class PostsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit(Post $post)
+    public function edit(Post $post, User $user)
     {
-        
-        return view('posts.edit-post')->with([
+        $this->authorize('edit', $post);
+
+        return view('posts.edit')->with([
             'post' => $post,
-            //'image' => $image
         ]);
     }
 
@@ -157,6 +158,8 @@ class PostsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('update', Post::class);
+
         $post = Post::find($id);
 
         $post->update($request->all());
@@ -174,8 +177,10 @@ class PostsController extends Controller
      */
     public function destroy(Post $post)
     {
-        $post->delete();
+        $this->authorize('delete', Post::class);
 
+        $post->delete();
+        
         return redirect()->route('posts.index');
     }
 
@@ -185,10 +190,11 @@ class PostsController extends Controller
      */
     public function restore($slug)
     {
+        $this->authorize('restore', Post::class);
+
         $post = Post::withTrashed()->where('slug', $slug)->first();
-
         $post->restore();
-
+        
         return redirect()->route('posts.show', $post);
     }
 }
